+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### Enhancement
 - Tests added to check that variables of type choices have specified a domain
 - Tests added to make sure the JSON object `id` matches the variable file name
+- Use of Ruby 2.1.1 is now recommended
 
 ## 0.4.1 (August 16, 2013)